+++ src/components/Loader/Loader.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+import * as S from './styles'
+
+// export interface LoaderProps {
+//   loading: boolean
+// }
+
+const Loader = () => {
+  return <S.Loader>Loading...</S.Loader>
+}
+
+export default Loader