+++ src/components/Loader/styles.ts
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+
+const Loader = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  height: 100%;
+`
+
+export { Loader }