+++ src/containers/Feedback/Feedback.tsx
@@ -8,6 +8,7 @@ import { getUserSession } from 'utils/googleAuth'
 import Comments from 'containers/Comments'
 import { readComments } from 'utils/firebase'
 import { sanitizeCommentsFromFirebase, EVENT_ID } from 'utils/helpers'
+import Loader from 'components/Loader'
 
 export interface FeedbackProps {
   active: boolean
@@ -34,6 +35,8 @@ class Feedback extends React.Component<FeedbackProps> {
     signInError: false,
     commentsError: false,
     comments: [],
+    loading: false,
+    noUserFound: false,
   }
 
   componentDidMount() {
@@ -51,28 +54,53 @@ class Feedback extends React.Component<FeedbackProps> {
     // so let's first check if there is
     // an active user session avail
 
-    firebaseInitialized && this.handleGetFirebaseUser()
+    // firebaseInitialized && this.handleGetUserSession()
 
     // run method to check for signed user
     // this will get the user data after a
     // successful Google sign in from the re-direct
     // and set the data in component state
     // if it fails, we'll trigger an error message
-    firebaseInitialized && this.handleGetFirebaseUser()
+    // firebaseInitialized && this.handleGetFirebaseUser()
 
-    firebaseInitialized && this.handleGetComments()
+    // firebaseInitialized && this.handleGetComments()
   }
 
   componentDidUpdate() {
-    const { displayName } = this.state
+    const { displayName, comments, loading, noUserFound } = this.state
+
     const firebaseInitialized = firebase.apps.length > 0
 
-    if (!displayName && firebaseInitialized) {
-      this.handleGetFirebaseUser()
+    if (firebaseInitialized) {
+      if (!displayName && !noUserFound && !loading) {
+        return this.handleGetFirebaseUser()
+      }
+    }
+
+    // subscribe to get comments
+    if (displayName) {
+      const commentsRef = firebase.database().ref('comments/')
+
+      commentsRef.on('value', (snapshot) => {
+        const commentsFromFB = sanitizeCommentsFromFirebase(snapshot.val())
+
+        if (comments.length !== commentsFromFB.length) {
+          this.setState(() => ({ comments: commentsFromFB }))
+        }
+      })
     }
   }
 
   handleGetFirebaseUser = () => {
+    const { active } = this.props
+    const { displayName } = this.state
+
+    if (displayName && !active) {
+      return null
+    }
+
+    this.setState(() => ({ loading: true, noUserFound: true }))
+
     firebase.auth().onAuthStateChanged((user) => {
       if (user) {
         const { displayName, email, photoURL, refreshToken } = user
@@ -82,32 +110,42 @@ class Feedback extends React.Component<FeedbackProps> {
           email,
           refreshToken,
           avatar: photoURL,
+          loading: false,
+          noUserFound: false,
         }))
+      }
+
+      if (!user) {
+        this.setState(() => ({ loading: false, noUserFound: true }))
       }
     })
   }
 
   handleGetUserSession = () => {
+    const { loading, displayName } = this.state
+
+    if (displayName) {
+      return null
+    }
+
+    !loading && this.setState(() => ({ loading: true }))
+
     getUserSession()
       .then((response) => {
         if (!response.error) {
           const { credential, displayName, email, avatar } = response
-          console.log({
-            credential,
-            displayName,
-            email,
-            avatar,
-          })
+
           return this.setState({
             credential,
             displayName,
             email,
             avatar,
+            loading: false,
           })
         }
       })
       .catch((err) => {
-        return this.setState({ signInError: true })
+        return this.setState({ signInError: true, loading: false })
       })
   }
 
@@ -127,7 +165,7 @@ class Feedback extends React.Component<FeedbackProps> {
 
   render() {
     const { active } = this.props
-    const { email, avatar, comments } = this.state
+    const { email, avatar, comments, loading } = this.state
 
     // addon not focused
     if (!active) {
@@ -135,6 +173,10 @@ class Feedback extends React.Component<FeedbackProps> {
     }
 
     if (active) {
+      if (loading) {
+        return <Loader />
+      }
+
       // if user not signed in
       if (!email) {
         return <Login />